Create a file for RoomAccessRule enum
